Add Item methods _resolve and _resolve_dict
These methods intend to replace current blender() logic in a
per-property level. The _resolve() method is used for non-dict
properties (booleans, strings, lists), _resolve_dict() is a special-case
merger for dictionaries.

Also in order to support '!foo' syntax the dict_removals() function is
now split in two:

  - dict_annihilate
  - dict_removals

The new function dict_annihilate takes a single dict as an argument and
is used by _resove_dict().