refactor: Remove unreliable caching
This will remove our caching feature due to the fact that it does not
work reliably and causes more problems than benefits. @schoolguy and I
discovered this while testing #2684.

Nevertheless in the future we should implement a new caching feature to
improve disk I/O.

Some calls to serialize() were missing and therefore the updates
regarding the children did not get updated properly.

Signed-off-by: Dominik Gedon <dgedon@suse.de>