Adjustments to cobbler-settings validate
This will expend the validate command of cobbler-settings in order to
tell the user what is wrong with the given settings files. Furthermore
there is a new `--version` option for validate where the user can
provide a version to validate the given settings file against. The
default is the most recent version of Cobbler.

Signed-off-by: Dominik Gedon <dgedon@suse.de>