Add types to various files
This PR adds basic Python type annotations, so we do get into a state where we
know what method calls require what types. This is important to have a
well-known state. This means that after the top level user API we can rely on
the types we are using and don't have to paranoidly check them at all times.

Apply suggestions by @agraul

Co-authored-by: Alexander Graul <agraul@suse.com>