Version bump to V3.3.1
This increments the version to 3.3.1 at all places and adds a new
migration for the settings.

This commit contains a fix so that we round the floating point
representation of the version in a way we hope it returns the correct
result. Since floats are just an approximation of the real result, this
is the best we can do for now. The solution we will strive for is to
remove the float representation of the number in the next bigger
release.