Add system tests
This commit adds simple system testing framework for Cobbler as well as
a bunch of essential tests. The framework is mostly intended for
black-box testing, meaning that we should be able to verify that things
like distro importing or OS installation actually work.

The interface is pretty self-explanatory:

  system-tests/run-tests  <- Use it to run tests
  system-tests/prelude    <- Helpers for tests
  system-tests/tests      <- A directory full of tests

By default run-script will run the most important and not-so-much
time-consuming tests.

You can also specify the tests you want to run; the usual globbing rules
apply:

  system-tests/run-tests basic-distro-rename *-distro-*

The test results can be found here:

  system-tests/workspace
  system-tests/workspace/<test-name>/output