system-tests: Check distribution importing
This commit introduces a generic test for distribution importing. During
the signature scans Cobbler does not care about the file contents and
only checks the file names, giving us an opportunity to cheat.

Instead of using a full-blown ISO image we use a replica. With the help
of the new script list-iso the directory structure of any installation
media can copied and then recreated using empty files during the test:

  system-tests/scripts/list-iso /tmp/FooBar-42-x86_64.iso \
          >system-tests/listings/foobar-42/FooBar-42-x86_64.iso

The test itself consists of the two files:

  import-@
  import-redhat-rhel8

The import-@ file contains the actual test body, while
import-redhat-rhel8 is a symlink. The test strips 'import-' from it's
name and uses the resulting string to lookup the directory with ISO
listings.

Thus, adding a new breed/version is just a matter of creating another
symlink. The import-@ file itself is ignored by run-tests.