system-tests: Check Grub autoinstall
This commit introduces a test that adds a dummy OS distribution and runs
QEMU that should be "provisioned".

This OS is actually a Linux kernel and the special busybox-based
initramfs that sends DHCP request, obtains the installation script from
Cobbler, notifies it about succesfull installation and reboots into the
disk image that shuts the VM down.

Easy.

However, since GitHub forbids Git LFS on public mirrors I'm currently
not pushing the kernel and initramfs. We will have to do this through
the master branch.