Fixup the item rename
During the rename operation we work with the object itself and thus
the reference to it in the dict. We can break down a rename operation
into the following steps:

    1. Remove object from dict (only needed because name is key)
    2. Rename the object (we have it in memory still)
    3. Readd the object to the dict
    4. Sync the changes in all related files

Steps 1 and 3 contain the removal and serialization of the object to
disk.

Step 4 should be done implicitly when we change the references to the
children which is also only required because we use the name instead
of the uuid.