docker: add build of debs for Debian 11 Bullseye
This commit adds a Dockerfile to build the debs for Debian 11. For the
build to succeed, update the file cobbler.spec to define the correct
package name in createrepo_pkg since Debian 11 moved to the C
implementation of createrepo.

Update the installation guide to provide the build and install command
for Debian 11. Also add Debian 11 to the Makefile and GH Action jobs.