mkloaders: Fixup syslinux for Debian based systems
Same as the commit before: We don't want to limit ourselves to the
runtime of (open)SUSE and/or CentOS but do want to be able to run
also on Debian based distros. Thus we make the directories for
pxelinux and memdisk configurable because they are packed separately
on Debian based distros.