Fix: Respect auto_migrate_settings
The root cause of ignoring this before is a design error in the
system how I imagined it:

Currently, we pass as a default value True, so we do migrate
per-default. So far this is fine. The problem is in that chain that
the value from the settings.yaml file is read but not taken into
consideration, thus the value from the daemon is used as the only
source of truth. Since we did add this key to the settings, we do
want to respect the value from the settings.

I define the desired behavior as:

- Default value is (if option not present anywhere): Do execute the
  auto-migration.
- Daemon Flag overpowers the value in settings.yaml.

Small hickup regarding the tests: Since we have a class wide
variable for sharing the state of the CobblerAPI we have to "guess"
the private name of those two variables and then reset them to have
a clean state.

Second small hickup: We do write to the settings file and thus the
test is not able to function identical every time. In theory, we
could solve this through mocking, but instead I decided to fix this
with a fixture which saves the state of the settings file and writes
it back after the test.

Co-authored-by: Pablo Suárez Hernández <psuarezhernandez@suse.com>