Enums: Create general str to enum converter
This commit deduplicates code and introduces a new method called
"to_enum(value)" which accepts str or enum type values and converts
them to enum type values.

Tests were moved from "validate_test.py" to "enums_test.py" and more
cases were added to check that the "<<inherits>>" logic works.

This effort is needed because we do need to stay Python 3.6
compatible. If we had been able to use Python 3.10, then we could
have just used StrEnum.

The second reason we need this, is that we don't know what type the
value might have. From the CLI it might be a str, from the demon
itself it might be a typed value already.