Security: Validate the data before logging it
This is required so that we don't create an opportunity for log file
pollution.

Scenario is the following: You issue an arbitrary HTTP request that
a system has finished installing. For this you don't need to be
authenticated and any valid str would be passed to the pre & post
install triggers. The validation now introduced will secure the
triggers.