Systems: Re-enable the *_interface calls
This reintroduces modify_interface to the System object. It is a
compatibility function. The way to do this now is:

    system.interfaces[<name>].property = value

The old way was:

    system.modify_interface({"property-<name>": value})

However since the new way is not yet supported by all consumers of
the API this compatibility layer.

This commit as well reintroduces "rename_interface". The usage didn't
change for XML-RPC API consumers, however it does for Python API
users. An example for XML-RPC:

     remote.modify_system(
            system,
            "rename_interface",
            {"interface": "eth0", "rename_interface": "eth_new"},
            token,
        )

Lastly this commit also reintroduces "delete_interface". The usage
didn't change for XML-RPC API users but it did for Python API users.
An example for the XML-RPC can be viewed below:

     remote.modify_system(
            system, "delete_interface", {"interface": "eth0"}, token
        )