Utils: Improve error handling of server_restart
Currently, when a service is not correctly started, supervisor will
return an exception and Cobbler crashes. Since we expect that during
a restart the service is already running we need to catch that.

This commit fixes this behavior and instead logs this to the Cobbler
log as an error.

This commit also improves the formatting of the service restart
message.