Tests: Switch from XML-RPC to Python
This commit switches from XML-RPC calls to Python to remove the
potential failure point of the network. Additionally, this should
enable the coverage reporter to track which lines are being called
in the XML-RPC API during testing.

The switch revealed that many of our tests do not have a correct
cleanup fixture. This commit thus contains also efforts to make
the tests more idempotent from each other.