Settings: Add proxies key
This commit introduces first-level support for Uyuni Proxies. The only thing
missing for this was to allow a new settings key called "proxies". The name is
not perfect for the moment, but changing the name to a more fitting one will be
done in a future version to make the changes minimal for both Uyuni and Cobbler
users alike.

More explanation for this feature of Uyuni can be found the linked document in
"cobbler-conf.rst".

Co-authored-by: Dominik Gedon <dgedon@suse.de>