System: Move profile and image validation
Currently we were able to get into an inconsistent data state through using an
image as a parent. To fix this we moved the name validation to the top of the
function.

On top of that we had to add validation on the image property because it was
only present in the profile property.