Add get_item_resolved_value
This commit adds a new method to allow querying an objects final values in case
the attribute is affected by inheritance. This allows non Python clients to
work with those values.

In the context of kernel_options this does not add additional options since these
are dependent on the scenario they are being used in.