Obsolete modules.conf
This commit is containing the obsoletion of the file
"/etc/cobbler/modules.conf". The file is important to decide which module
is being loaded for the different parts of DHCP, DNS and TFTP management, as
well as serialization and deserialization of the database.

The functionality of the file is now included in our main settings file. To
make this possible it was required to refactor the module loader to an object
and pass the API to the serializer.

In contrast to the behavior of the "modules.conf" the new settings are
optional and only need to be overridden. The schema is validated as usually.

This commit also changes the message the logger returns to dot notation.