XML-RPC API: Fix complex objects are return value
This bug is appearing in "get_item_resolved_value" and could be triggered for
example when "arch" was asked with a distro. The fix is to catch all types in
"remote.py" and convert them on a per-type basis.

This commit also adds tests for the known cases that would cause trouble in the
past.