Do not validate costly network interface attributes when data is deserialized from disk
Since the new getter/setter concept in items/*, data is validated
in the setter function.
Therefore validation also happens, when data is deserialized from disk.
Even the data there already had been validated.

This patch introduces a hackish quick fix and introduces the state
is_deserialize
in items/system.py and items/item.py
When the object is created and data gets filled from deserialization,
the most costy validations (iterate over all other systems and check whether
dns_name, ipv4, ipv6 or mac address already exists) are ignored and the
attribute is directly assigned.

This decreases cobbler start time with about 500 systems from more than 4 minutes
down to about 4 seconds:
[Daemon] 2021-11-17T19:30:48 - INFO | Cobbler startup completed  in 265.8612322807312 seconds
[Daemon] 2021-11-17T20:05:09 - INFO | Cobbler startup completed  in 3.5047478675842285 seconds

Be aware that the entry point is in:
cobbler/cobbler_collections/systems.py
+        new_system.is_deserialize = True

The code in cobbler/items/system.py:
deserialize()
self.is_deserialize = True
is not used, but if it should be used in the future this change is needed there (from code):

IMO the correct longterm approach could be to pass the dict
as kwargs in the constructor when the object is created via deserialization.
Attribute assignment via constructor should be a nice way to
get around attribute/setter validations on purpose:

E.g.:
instead of:
         new_system = system.System(api)
         new_system.from_dict(item_dict)
do:
         new_system = system.System(api, **item_dict)

and in items/system.py the constructor would need to be enhanced like:
cobbler/items/system.py:
         :param api: The Cobbler API
         """
         super().__init__(api, *args, **kwargs)
-        self._interfaces: Dict[str, NetworkInterface] = {"default": NetworkInterface(api)}
-        self._ipv6_autoconfiguration = False
+        self._interfaces = interfaces if interfaces else self._interfaces = {"default": NetworkInterface(api)}
+        self._ipv6_autoconfiguration if ipv6_autoconfiguration else self._ipv6_autoconfiguration = False

.. to be discussed