API: Improvements on missing directories
This commit improves the error handling of missing directories during startup of
the daemon. Since this is located in the CobblerAPI class the code will also
be executed when the PythonAPI is used. There is currently a mechanism lacking
that is able to distinguish automatically between the daemon and the API Client
usage.