Consider case of next_server being a hostname during migration
If "next_server" is a hostname and not an IPv4 nor IPv6 address
then we try to resolve this hostname to fill the corresponding
new "next_server_v4" and "next_server_v6" settings.

If resolution does not work for any of those settings, and error
is printed and migration of "next_server" is not performed.