Prevent issues in case request.node.originalname is not available
We have noticed that, under some situations, there is no "originalname"
attribute set for "request.node", making pytest to crash.

It is not yet clear if this error is caused by different pytest version,
or maybe only under certaing setup conditions, so for now, we just make
our code to avoid the possible exception to happen under these
conditions.