Avoid possible override of existing values
When running the migration of collections from older Cobbler 2
collections to newer to 3.3.0, the mechanism for adding new items
to collections (or renaming them) is not taking into consideration
that those items may be existing already and may have different values
than the default ones to add.

We have detected this happening in distros where "boot_loaders" attribute
was already existing because of any reason (i.a. retriggering migration),
with "boot_loaders" set to "<<inherit>>", then after the migration runs
the "boot_loaders" will end up with default value "grub", instead of
preserving the existing value. This makes then profiles for this distro
to not being available in the generated PXE menus.