Skip wrong collections instead of crashing
This commit makes Cobbler service more tolerant to any unexpected
issue when loading collections. Instead of making Cobbler to crash,
this commit make it to skip collections that cannot be loaded and keep
the service running.

This way, we are not blocking other collections and particularly the
entire Cobbler service to work in case of a problem with a particular
collection.