Fix param-file for s390 systems when long parameters
The previous code was not handling well the cases where kernel option is
longer than 80 characters.

parm file format is fixed to 80 chars per line.
All the lines are concatenated without spaces when being passed to the kernel.

Recommendation: one parameter per line (ending with whitespace)

NOTE: If a parameter is too long to fit into the 80 characters limit it can simply
be continued in the first column of the next line.

https://www.debian.org/releases/stable/s390x/ch05s01.en.html
https://documentation.suse.com/sles/15-SP1/html/SLES-all/cha-zseries.html#sec-appdendix-parm-examples
https://wiki.ubuntu.com/S390X/InstallationGuide