Fix filename inheritence
The check for parent in profile item setter is wrong:
profile always has a parent (distro).

It has to be checked whether the parent is (not) of type profile
anymore (is_subobject). In this case we know we are at the end
of inheritance hierarchy and we have to set filename=""