Managers: Split into DHCP, DNS and TFTP managers
Since we have different types of managers we need to have abstract sub classes for the different kinds of managers. The current implementation was not able to correctly type check and seperate them by type. The new manager types are DHCP, TFTP and DNS. The methods are defined in a way that the current implementation doesn't need to be adjusted. Further cleanup is still required.