Fixes problem with efi netboot
Had problems with netbooting efi systems with grub. The last entry
from the $arch_menu.cfg was booted, even if a system entry with MAC
existed. It was also loaded.
Changing the order of loading the files made the system specific entry
the last one in the list and it was booted.