Make the CNAME rendering to actually be aligned
This is purely a cosmetic change, in how the CNAME aliases are aligned
in the zone files. Previously a fixed number of spaces was inserted
between the tokens in the zone file, making it look unappealing. The
same column-alignment logic that is already used for the A, AAAA and
PTR records is now applied to the CNAME section too.

Fixes #4070