Tests: Force garbage collect for get_autoinstall
Due to reasons beyond my understanding, the tests ran out of memory on
many machines. This meant that the performance testsuite easily ate up
my 32GB of laptop RAM. After a little bit of investigation it showed
that the root-cause were the get_autoinstall tests. To prevent this a
manual garbage collection cycle seems to do the trick. Since the time it
takes to execute the garbage collection is rather constant, this only
adds a static number of time on top of the time needed to generate the
autoinstallation template.