Get rid of data_files in setup.py
This introduces a new command for "cobblerd" which is called "setup".
The new command is replacing the work that setuptools previously did and
installing the files in the file system of the host.

The "data_files" attribute in "setup.py" cannot be used as its usage is
discouraged since 2021. Furthermore the introduction of the command
allows cobbler to be installed via "pip".