snippet 'keep_files': global variable 'keys_found' should be per-service
Snippet `keep_files` attempts to preserve keys from potentially multiple services:
currently 'ssh', 'cfengine', 'rhn', 'puppet'.

Each per-service key-search of the old OS sets a variable `keys_found` to `yes` or `no`,
partly with the intention of guiding the behaviour of the subsequent `restore_keys` function.

The first act of each per-service call to `search_for_keys` sets `keys_found` to `no`.
In particular, if the final service examined results in `no`, then its final value is `no`.

That single variable is, in effect, global across all the services.

The first act of each per-service `restore_keys` is to examine that single variable.
So if the final service that had been examined had resulted in `no`, then no keys
are restored for any of the services.

This change converts that single variable into a per-service array of variables
for the benefit of `restore_keys`.