Items: Use uid as primary key
This commit is using the UID instead of the name as the primary
reference between items. This should enable collections to allow
duplicate names if desired. Furthermore, this simplifies the rename
operation as no file needs to be renamed anymore and no references
inside other items need to be updated.

This commit further enhances the docstrings of various tests and methods
that were encountered during the process of moving from names to UIDs.